requirements.txt

streamlit==1.28.0
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
plotly==5.15.0
openpyxl==3.1.2
joblib==1.3.0
python-dotenv==1.0.0
statsmodels==0.14.0